Braids

Description

Documentation

data Braid

A Braid is a width and word. Integers in the word represent Artin generators and their invereses. E.g. [1,3,-2] represents the word sigma_1sigma_3sigma_2^{-1}.

Constructors

Braid

Fields

braidWidth :: Int

braidWord :: [Int]

data Node

A braid can also be written as a collection of joined nodes. See knotatlas for more info on Cross and Join.

Constructors

Cross Int Int Int Int

Join Int Int

type PD = [Node]

A PD (Planar Diagram) is a collection of Nodes.

type Resolution = [Int]

A Resolution is a collection of integers. These should all be 0 or 1 and should probably be represented by Bools instead of Ints.

data Diagram

A Diagram has a resolution and a set of components.

Constructors

Diagram

Fields

resolution' :: Resolution

components' :: [[Int]]

parse :: [String] -> (Braid, Int)

Parse command line input into a braid

ngkRight :: (Int, Int) -> Braid

These functions construct a family of braids from Ng and Khandawit's paper:

ngkLeft :: (Int, Int) -> Braid

braidToPD :: Braid -> PD

Parse a Braid into a PD.

crossingToNodes :: Int -> Int -> Int -> [Node]

Parses a crossing into a Node.

isCross :: Node -> Bool

allRes :: PD -> [Resolution]

resolve :: PD -> Resolution -> PD

Takes a PD and a Resolution and returns the resolved PD. Note that the output is always a list of joins.

resolutions :: PD -> [PD]

Computes all resolutions of a PD as a list of PDs.

resolutionToComponents :: PD -> [[Int]]

cubeOfResolutions :: PD -> [Diagram]

Takes a PD and returns a list of all the Diagrams of the resolutions.

braidCube :: Braid -> [Diagram]

The total cube of resolutions for a braid.