Complex

Documentation

data Generator

A Generator has a resolution, a list of components, a list of signs (all 1 or -1), and a k-grading. The k-grading is special in that it is determined by the other three parts, but this is not indicated in the data structure.

Constructors

Generator

Fields

resolution :: Resolution

components :: [[Int]]

signs :: [Int]

kgrade :: Int

type Morphisms = Map Generator (Set Generator)

Morphisms is a map from generators to sets of generators

data Cancellations

Cancellations are Morphisms, but they have a special composition

Constructors

Cancellations (Map Generator (Set Generator))

consume :: Ord k => Map k (Set k) -> Map k (Set k) -> Map k (Set k)

consumeKey :: Ord k => Map k (Set k) -> (k, Set k) -> Map k (Set k)

addMod2 :: (Eq a, Ord a) => a -> Set a -> Set a

The next three items implement mod 2 addition at the level of sets and Morphisms

addMod2Set :: (Eq a, Ord a) => Set a -> Set a -> Set a

addMod2Map :: Map Generator (Set Generator) -> Map Generator (Set Generator) -> Map Generator (Set Generator)

simplifyEdgeGraph :: (Generator, Generator) -> Morphisms -> Morphisms

Return the complex simplified at (g,g')